adjust director
