+++ code/app/commodity/svc/service_context.go
@@ -1,0 +1,13 @@
+package svc
+
+import "gorm.io/gorm"
+
+type ServiceContext struct {
+	Db *gorm.DB
+}
+
+func NewServiceContext() *ServiceContext {
+	return &ServiceContext{
+		Db: NewDBService(),
+	}
+}
